messed up asynch

+++ GameHubApp/app/src/main/java/com/gamehub/myapplication/games.java
@@ -13,9 +13,9 @@ public class games {
     private Double gameID;
     private String title;
     private String publisher;
-    private Integer rating;
+    private Double rating;
     // Should I have a slot for avg rating and also make a list of all of its ratings maybe mapped to user id so you cant have more than one vote
-    public games(Double gameID, String title, String publisher, Integer rating){
+    public games(Double gameID, String title, String publisher, Double rating){
         this.gameID = gameID;
         this.title = title;
         this.publisher = publisher;
@@ -30,7 +30,7 @@ public class games {
     public void setPublisher(String publisher){
         this.publisher = publisher;
     }
-    public void setRating(Integer rating){
+    public void setRating(Double rating){
         this.rating = rating;
     }
     public Double getID(){
@@ -42,7 +42,7 @@ public class games {
     public String getPublisher(){
         return this.publisher;
     }
-    public Integer getRating(){
+    public Double getRating(){
         return this.rating;
     }
 }
